Add files via upload
I was able to get everything to work correctly until the last video (Twitter 2: Episode 4 (Stretch) - Persistence.) I watched the video again multiple times to ensure that I have no errors and that everything was correct. I did not find anything I did differently from the video, but my app shuts down each time as I try to open it.
Before it started to crash, I was able to create my own post and post it on a time line. As show in the gif, those post were done through the emulator and posted in real time on twitter.